first part of splitting modules
