+++ EventFilter/L1TRawToDigi/src/OmtfRpcLinkMap.cc
@@ -1,0 +1,1 @@
+#include "EventFilter/L1TRawToDigi/interface/OmtfRpcLinkMap.h"
